seekbar is able to pass values to c# now
-the new time was a float not an int value

+++ Jukebox/JukeboxMediaManager.cs
@@ -218,7 +218,7 @@ namespace Jukebox.Library
                 
             });
         }
-        public int Seek(int seekValue)
+        public int Seek(float seekValue)
         {
             Application.Current.Dispatcher.BeginInvokeOnMainThread(async () =>
             {
